Introduce softmax layer
Previously, the model returned a 1 x 128 tensor, which does not tell us
anything about the predicted word. We therefore try to map this through
a linear layer from 128 to vocab size, then run this layer through
softmax. I have no idea if this is the correct way to handle the output.

It also turns out that if the learning rate is too high, then the
gradient eventually explodes. We therefore reduce the learning rate.